Move neighborlist tests to separate file
